Make the matched-factor a tunable parameter of the SkeletonEndPointTracker.
This parameter specifies by how much to decrease the cost of
linking two skeleton end-points that connect to the same junction track.
For instance if the distance-based cost to link e1 and e2 is 250 but they
are connected to the same junction track, as tracked by the junction
track ID stored in the {@link JunctionIDAnalyzerFactory#FEATURE}, the
actual linking cost will be 250 divided by the factor stored in this
parameter.