WIP: Robust branch length calculation.
Now I have something that can path incorrect branch length when the
junction it connects to changes.
What we do: select the main junction IDs connected to an end-point
tracks, and see everytime it does not connect to it if we can find
it further in the skeleton graph by BFS. If we can, we compute the
new branch legnth by summing each branch length.
This is cool, and it works.
Now we need to determine what is the best among all possible
junction id candidate.