Manual Track feature to store the number of incorrect junction IDs.
This track feature stores the number of incorrect junction IDs its
end-point spots link to.

In ideal situation a dendrite track is made of a collection of
end-points in the neuron skeleton that always link to the same
junction track ID. Because of incorrect tracking either for the
end-points or for the junctions, it might not be the case. This
feature stores the number of spots in the track that do not link
to the main junction track ID. A value of 0 means that the end-point
track has no defects.