Algorithms that performs the dendrite track analysis.
For now it computes the branch length.
Its main interest is that it tries to patch the correct branch
length when the end-point track does not link to the same junction
track ID over time.
For those points it tries to retrieve the junction ID by a BFS search
on the skeleton.
This is important because of the other dendrite dynamics that might
appear and disappear along the branch we are trying to calculate the
length of.
So this algorithm provides the following features:
SPOTS:
- branch length
- junction id
TRACKS:
- number of spots in the track with 'incorrect' junction id.
A value of 0 means that the full track has a correct branch length.