Work around strange TrackMate bug: tracking spots with identical positions.
It turns out TrackMate cannot track reliably spots that have exactly
the same position over time. It happens in our case since our positions
are integers.
A work around is to add a small random value to their position and it
works.

This bug can probably be reproduced in TrackMate when disabling the
sub-pixel accuracy flag in the LoGDetector....