POM: fix project attributes
The XML schema for Maven POMs is located at:

  https://maven.apache.org/xsd/maven-4.0.0.xsd

Its XML namespace is the string:

  http://maven.apache.org/POM/4.0.0

So that exact string must be the value of xmlns. It must also
match the first half of xsi:schemaLocation, which maps that
namespace to an actual URL online where the schema resides.
Otherwise, the document is not a Maven POM.

Similarly, the xmlns:xsi attribute of an XML document declaring a
particular schema should always use the string identifier:

  http://www.w3.org/2001/XMLSchema-instance

because that's the namespace identifier for instances of an XML schema.

For details, see the specification at: https://www.w3.org/TR/xmlschema-1/